+++ contributions/catalog/controllers/contribute.py
@@ -4,23 +4,21 @@ import traceback
 # from app import app
 import requests
 from flask import (
-    Blueprint, render_template, request, session, redirect, url_for
+    Blueprint, render_template, request, session, redirect
 )
-from requests_oauthlib import OAuth2Session
-
 from controllers.config import Config as cfg
-from controllers.config import Config
+from requests_oauthlib import OAuth2Session
 from models.contribution_utilities import to_contribution
+from utils import jsonutil
 
 bp = Blueprint('contribute', __name__, url_prefix='/contribute')
 
 @bp.route('/', methods=['GET', 'POST'])
+
 def home():
+    print("homepage.")
     if request.method == 'POST' and request.validate_on_submit():
-        # print("searching...")
         result = request.form.to_dict(flat=False)
-        # print(result)
-        # search(result)
     if "name" in session:
         return render_template('contribute/home.html', user=session["name"])
     else:
@@ -31,7 +29,7 @@ def login():
     """Step 1: Get the user identify for authentication.
     """
     # print("Step 1: User Authorization")
-    github = OAuth2Session(cfg.CLIENT_ID)
+    github = OAuth2Session(cfg.GITHUB_CLIENT_ID)
     authorization_url, state = github.authorization_url(cfg.AUTHORIZATION_BASE_URL)
 
     # State is used to prevent CSRF.
@@ -48,9 +46,7 @@ def logout():
 def result():
     if request.method == 'POST':
         # result = request.form
-        # print("searching...")
         result = request.form.to_dict()
-        # print(result)
         query = result['search']
         search(query)
         return render_template("contribute/results.html", user=session["name"], result=result)
@@ -63,21 +59,23 @@ def create():
         # result = dict((key, request.form.getlist(key) if len(request.form.getlist(key)) > 1 else request.form.getlist(key)[0]) for key in request.form.keys())
 
         contribution = to_contribution(result)
-        # print(contribution)
+
+        # add contributionAdmins to the json_contiubtion
+        contribution = jsonutil.add_contribution_admins(contribution)
+
         json_contribution = json.dumps(contribution, indent=4)
-        # print(json_contribution)
         response, s = post(json_contribution)
         if response:
-            if "name" in session:
-                return render_template('contribute/submitted.html', user=session["name"])
-            else:
-                return render_template('contribute/submitted.html')
-        elif not response:
-            if "name" in session:
-                return render_template('contribute/error.html', user=session["name"], error_msg=s)
-            else:
-                return render_template('contribute/error.html', error_msg=s)
-
+            if response:
+                if "name" in session:
+                    return render_template('contribute/submitted.html', user=session["name"])
+                else:
+                    return render_template('contribute/submitted.html')
+            elif not response:
+                if "name" in session:
+                    return render_template('contribute/error.html', user=session["name"], error_msg=s)
+                else:
+                    return render_template('contribute/error.html', error_msg=s)
     return render_template('contribute/contribute.html', user=session["name"])
 
 @bp.errorhandler(404)
@@ -99,20 +97,20 @@ def search_results(search):
 def post(json_data):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + Config.AUTHENTICATION_TOKEN
+        'Authorization': 'Bearer ' + session['oauth_token']['access_token']
     }
     try:
         # Setting up post request
-        result = requests.post(Config.CONTRIBUTION_BUILDING_BLOCK_URL,
+        result = requests.post(cfg.CONTRIBUTION_BUILDING_BLOCK_URL,
                                headers=headers,
                                data=json_data)
 
         if result.status_code != 200:
-            # print("post method fails".format(json_data))
-            # print("with error code:", result.status_code)
+            print("post method fails".format(json_data))
+            print("with error code:", result.status_code)
             return False, str("post method fails with error: ")+str(result.status_code)
         else:
-            # print("posted ok.".format(json_data))
+            print("posted ok.".format(json_data))
             return True, str("post success!")
 
     except Exception:
@@ -125,24 +123,24 @@ def post(json_data):
 def search(input_data):
     headers = {
         'Content-Type': 'application/json',
-        'Authorization': 'Bearer ' + Config.AUTHENTICATION_TOKEN
+        'Authorization': 'Bearer ' + cfg.AUTHENTICATION_TOKEN
     }
 
     try:
         # Setting up post request
         if not input_data or len(input_data) == 0:
-            result = requests.get(Config.CONTRIBUTION_BUILDING_BLOCK_URL + "/",
+            result = requests.get(cfg.CONTRIBUTION_BUILDING_BLOCK_URL + "/",
                                   headers=headers)
         else:
-            result = requests.get(Config.CONTRIBUTION_BUILDING_BLOCK_URL + "/" + str(input_data),
+            result = requests.get(cfg.CONTRIBUTION_BUILDING_BLOCK_URL + "/" + str(input_data),
                                   headers=headers)
 
         if result.status_code != 200:
-            # print("post method fails".format(input_data))
-            # print("with error code:", result.status_code)
+            print("post method fails".format(input_data))
+            print("with error code:", result.status_code)
             return False
         else:
-            # print("posted ok.".format(input_data))
+            print("posted ok.".format(input_data))
             return True
 
     except Exception: